+++ lib/uploadcare/exception/auth_error.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Uploadcare
+  module Exception
+    # Invalid Auth configuration error
+    class AuthError < StandardError; end
+  end
+end