+++ lib/uploadcare/param/authentication_header.rb
@@ -11,6 +11,7 @@ module Uploadcare
     class AuthenticationHeader
       # @see https://uploadcare.com/docs/api_reference/rest/requests_auth/#auth-uploadcare
       def self.call(options = {})
+        validate_keys
         case Uploadcare.config.auth_type
         when 'Uploadcare'
           SecureAuthHeader.call(options)
@@ -19,6 +20,13 @@ module Uploadcare
         else
           raise ArgumentError, "Unknown auth_scheme: '#{Uploadcare.config.auth_type}'"
         end
+      end
+
+      private
+
+      def self.validate_keys
+        raise AuthError, "Public Key is blank." if Uploadcare.config.public_key.empty?
+        raise AuthError, "Secret Key is blank." if Uploadcare.config.secret_key.empty?
       end
     end
   end